@@ -5,7 +5,7 @@ author: Thai Dinh Le <Thai.Dinh.Le.pro@gmail.com>
 homepage: https://github.com/thai-dinh/adhoc_plugin
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.20.0"
 
 dependencies:
