@@ -8,12 +8,12 @@ part 'adhoc_device.g.dart';
 /// Bluetooth Low Energy, or both enabled.
 @JsonSerializable()
 class AdHocDevice {
-  bool directedConnected;
-  String label;
-  String address;
-  String _name;
-  String _mac;
-  int _type;
+  bool? directedConnected;
+  String? label;
+  String? address;
+  String? _name;
+  String? _mac;
+  int? _type;
 
   /// Creates an [AdHocDevice] object.
   ///
@@ -25,8 +25,8 @@ class AdHocDevice {
   /// If [address] is given, it is either an UUID in case of Bluetooth Low 
   /// Energy, or an IPv4 address in case of Wifi P2P.
   AdHocDevice({
-    this.directedConnected = false, this.label = '', String address = '', 
-    String name, String mac, int type = 2,
+    this.directedConnected = false, this.label = '', String? address = '', 
+    String? name, String? mac, int? type = 2,
   }) {
     this.label = checkString(label);
     this.address = checkString(address);
@@ -41,11 +41,11 @@ class AdHocDevice {
 
 /*------------------------------Getters & Setters-----------------------------*/
 
-  String get name => _name;
+  String? get name => _name;
 
-  String get mac => _mac;
+  String? get mac => _mac;
 
-  int get type => _type;
+  int? get type => _type;
 
 /*-------------------------------Public methods-------------------------------*/
 
