@@ -1,6 +1,6 @@
 class AdHocEvent {
   int type;
-  Object payload;
+  Object? payload;
 
   AdHocEvent(this.type, this.payload);
 }
