@@ -27,16 +27,16 @@ class AodvManager {
 
   final bool _verbose;
 
-  AodvHelper _aodvHelper;
-  HashMap<String, int> _mapDestSequenceNumber;
-  StreamController<AdHocEvent> _eventCtrl;
+  late AodvHelper _aodvHelper;
+  late HashMap<String?, int?> _mapDestSequenceNumber;
+  late StreamController<AdHocEvent> _eventCtrl;
 
-  String _ownMac;
-  String _ownName;
-  String _ownLabel;
-  int _ownSequenceNum;
-  DataLinkManager _datalinkManager;
-  MessageAdHoc _dataMessage;
+  String? _ownMac;
+  String? _ownName;
+  String? _ownLabel;
+  int? _ownSequenceNum;
+  DataLinkManager? _datalinkManager;
+  MessageAdHoc? _dataMessage;
 
   AodvManager(this._verbose, Config config) {
     this._aodvHelper = AodvHelper(_verbose);
@@ -53,17 +53,17 @@ class AodvManager {
 
 /*------------------------------Getters & Setters-----------------------------*/
 
-  String get label => _ownLabel;
+  String? get label => _ownLabel;
 
-  DataLinkManager get dataLinkManager => _datalinkManager;
+  DataLinkManager? get dataLinkManager => _datalinkManager;
 
   Stream<AdHocEvent> get eventStream => _eventCtrl.stream;
 
-  Stream<DiscoveryEvent> get discoveryStream => _datalinkManager.discoveryStream;
+  Stream<DiscoveryEvent> get discoveryStream => _datalinkManager!.discoveryStream;
 
 /*------------------------------Public methods-------------------------------*/
 
-  void sendMessageTo(Object pdu, String address) {
+  void sendMessageTo(Object pdu, String? address) {
     Header header = Header(
       messageType: AodvConstants.DATA,
       label: _ownLabel,
@@ -79,24 +79,24 @@ class AodvManager {
 /*-----------------------------Private methods-------------------------------*/
 
   void _initialize() {
-    _datalinkManager.eventStream.listen((AdHocEvent event) {
+    _datalinkManager!.eventStream.listen((AdHocEvent event) {
       switch (event.type) {
         case DatalinkConstants.BROKEN_LINK:
-          _brokenLinkDetected(event.payload);
+          _brokenLinkDetected(event.payload as String?);
           break;
 
         case DatalinkConstants.MESSAGE_EVENT:
-          _processAodvMsgReceived(event.payload);
+          _processAodvMsgReceived(event.payload as MessageAdHoc);
           break;
 
         case DatalinkConstants.DEVICE_INFO_BLE:
-          List<dynamic> info = event.payload;
+          List<dynamic> info = event.payload as List<dynamic>;
           _ownMac = info[0] as String;
           _ownName = info[1] as String;
           break;
 
         case DatalinkConstants.DEVICE_INFO_WIFI: // Always select this
-          List<dynamic> info = event.payload;
+          List<dynamic> info = event.payload as List<dynamic>;
           _ownMac = info[0] as String;
           _ownName = info[1] as String;
           break;
@@ -128,7 +128,7 @@ class AodvManager {
       display = true;
 
       buffer.write('--------Routing Table:--------\n');
-      for (MapEntry<String, EntryRoutingTable> entry in _aodvHelper.getEntrySet()) {
+      for (MapEntry<String?, EntryRoutingTable> entry in _aodvHelper.getEntrySet()) {
         buffer..write(entry.value.toString())..write('\n');
       }
     }
@@ -136,7 +136,7 @@ class AodvManager {
     if (_mapDestSequenceNumber.length > 0) {
       display = true;
       buffer.write('--------SequenceNumber:--------\n');
-      for (MapEntry<String, int> entry in _mapDestSequenceNumber.entries) {
+      for (MapEntry<String?, int?> entry in _mapDestSequenceNumber.entries) {
         buffer..write(entry.key)..write(' -> ')..write(entry.value.toString())..write('\n');
       }
     }
@@ -145,7 +145,7 @@ class AodvManager {
       print(buffer.toString());
   }
 
-  void _brokenLinkDetected(String remoteNode) {
+  void _brokenLinkDetected(String? remoteNode) {
     if (_aodvHelper.sizeRoutingTable() > 0) {
       if (_verbose) log(TAG, 'Send RRER');
       _sendRRER(remoteNode);
@@ -157,25 +157,25 @@ class AodvManager {
     }
   }
 
-  void _send(MessageAdHoc message, String address) {
-    if (_datalinkManager.isDirectNeighbors(address)) {
-      EntryRoutingTable destNext = _aodvHelper.getNextfromDest(address);
-      if (destNext != null && message.header.messageType == AodvConstants.DATA)
+  void _send(MessageAdHoc? message, String? address) {
+    if (_datalinkManager!.isDirectNeighbors(address)) {
+      EntryRoutingTable? destNext = _aodvHelper.getNextfromDest(address);
+      if (destNext != null && message!.header!.messageType == AodvConstants.DATA)
         destNext.updateDataPath(address);
       _sendDirect(message, address);
     } else if (_aodvHelper.containsDest(address)) {
-      EntryRoutingTable destNext = _aodvHelper.getNextfromDest(address);
+      EntryRoutingTable? destNext = _aodvHelper.getNextfromDest(address);
       if (destNext == null) {
         if (_verbose) log(TAG, 'No destNext found in the routing Table for $address');
       } else {
         if (_verbose) log(TAG, 'Routing table contains ${destNext.next}');
 
-        if (message.header.messageType == AodvConstants.DATA)
+        if (message!.header!.messageType == AodvConstants.DATA)
           destNext.updateDataPath(address);
 
         _sendDirect(message, destNext.next);
       }
-    } else if (message.header.messageType == AodvConstants.RERR) {
+    } else if (message!.header!.messageType == AodvConstants.RERR) {
       if (_verbose) log(TAG, 'RERR sent');
     } else {
       _dataMessage = message;
@@ -184,13 +184,13 @@ class AodvManager {
     }
   }
 
-  void _sendDirect(MessageAdHoc message, String address) {
+  void _sendDirect(MessageAdHoc? message, String? address) {
     if (_verbose) log(TAG, 'Send directly to $address');
 
-    _datalinkManager.sendMessage(message, address);
+    _datalinkManager!.sendMessage(message, address);
   }
 
-  void _startTimerRREQ(String destAddr, int retry, int time) {
+  void _startTimerRREQ(String? destAddr, int retry, int time) {
     if (_verbose) log(TAG, 'No connection to $destAddr -> send RREQ message');
 
     MessageAdHoc message = MessageAdHoc(
@@ -210,10 +210,10 @@ class AodvManager {
       )
     );
 
-    _datalinkManager.broadcast(message);
+    _datalinkManager!.broadcast(message);
 
     Timer(Duration(milliseconds: time), () {
-      EntryRoutingTable entry = _aodvHelper.getNextfromDest(destAddr);
+      EntryRoutingTable? entry = _aodvHelper.getNextfromDest(destAddr);
       if (entry == null) {
         if (retry == 0) {
           _eventCtrl.add(AdHocEvent(
@@ -229,45 +229,45 @@ class AodvManager {
     });
   }
 
-  List<String> _addPrecursors(String precursorName) {
-    return List<String>.empty(growable: true)..add(precursorName);
+  List<String?> _addPrecursors(String? precursorName) {
+    return List<String?>.empty(growable: true)..add(precursorName);
   }
 
-  int _getDestSequenceNumber(String dest) {
+  int? _getDestSequenceNumber(String? dest) {
     if (_mapDestSequenceNumber.containsKey(dest))
       return _mapDestSequenceNumber[dest];
     return AodvConstants.UNKNOWN_SEQUENCE_NUMBER;
   }
 
   void _getNextSequenceNumber() {
-    if (_ownSequenceNum < AodvConstants.MAX_VALID_SEQ_NUM) {
-      ++_ownSequenceNum;
+    if (_ownSequenceNum! < AodvConstants.MAX_VALID_SEQ_NUM) {
+      _ownSequenceNum = _ownSequenceNum! + 1;
     } else {
       _ownSequenceNum = AodvConstants.MIN_VALID_SEQ_NUM;
     }
   }
 
-  void _saveDestSequenceNumber(String dest, int seqNum) {
+  void _saveDestSequenceNumber(String? dest, int? seqNum) {
     _mapDestSequenceNumber.putIfAbsent(dest, () => seqNum);
   }
 
   void _processRREQ(MessageAdHoc message) {
-    RREQ rreq = RREQ.fromJson(message.pdu as Map);
-    int hop = rreq.hopCount;
-    String originateAddr = message.header.label;
+    RREQ rreq = RREQ.fromJson((message.pdu as Map) as Map<String, dynamic>);
+    int? hop = rreq.hopCount;
+    String? originateAddr = message.header!.label;
     if (_verbose) log(TAG, 'Received RREQ from $originateAddr');
 
-    if (rreq.destAddress.compareTo(_ownLabel) == 0) {
+    if (rreq.destAddress!.compareTo(_ownLabel!) == 0) {
       _saveDestSequenceNumber(rreq.originAddress, rreq.originSequenceNum);
 
       if (_verbose) log(TAG, '$_ownLabel is the destination (stop RREQ broadcast)');
 
-      EntryRoutingTable entry = _aodvHelper.addEntryRoutingTable(
+      EntryRoutingTable? entry = _aodvHelper.addEntryRoutingTable(
         rreq.originAddress, originateAddr, hop, rreq.originSequenceNum, AodvConstants.NO_LIFE_TIME, null
       );
 
       if (entry != null) {
-        if (rreq.destSequenceNum > _ownSequenceNum)
+        if (rreq.destSequenceNum! > _ownSequenceNum!)
           _getNextSequenceNumber();
 
         RREP rrep = RREP(
@@ -297,11 +297,11 @@ class AodvManager {
         // _timerFlushReverseRoute(rreq.originAddress, rreq.originSequenceNum);
       }
     } else if (_aodvHelper.containsDest(rreq.destAddress)) {
-      _sendRREP_GRATUITOUS(message.header.label, rreq);
+      _sendRREP_GRATUITOUS(message.header!.label, rreq);
     } else {
-      if (rreq.originAddress.compareTo(_ownLabel) == 0) {
+      if (rreq.originAddress!.compareTo(_ownLabel!) == 0) {
         if (_verbose) log(TAG, 'Reject own RREQ ${rreq.originAddress}');
-      } else if (_aodvHelper.addBroadcastId(rreq.originAddress, rreq.rreqId)) {
+      } else if (_aodvHelper.addBroadcastId(rreq.originAddress!, rreq.rreqId)) {
         rreq.incrementHopCount();
         message.header = Header(
           messageType: AodvConstants.RREQ, 
@@ -311,7 +311,7 @@ class AodvManager {
         );
         message.pdu = rreq;
 
-        _datalinkManager.broadcastExcept(message, originateAddr);
+        _datalinkManager!.broadcastExcept(message, originateAddr);
 
         _aodvHelper.addEntryRoutingTable(rreq.originAddress, originateAddr, hop, rreq.originSequenceNum, AodvConstants.NO_LIFE_TIME, null);
 
@@ -323,24 +323,24 @@ class AodvManager {
   }
 
   void _processRREP(MessageAdHoc message) {
-    RREP rrep = RREP.fromJson(message.pdu as Map);
-    int hopRcv = rrep.hopCount;
-    String nextHop = message.header.label;
+    RREP rrep = RREP.fromJson((message.pdu as Map) as Map<String, dynamic>);
+    int? hopRcv = rrep.hopCount;
+    String? nextHop = message.header!.label;
 
     if (_verbose) log(TAG, 'Received RREP from $nextHop');
 
-    if (rrep.destAddress.compareTo(_ownLabel) == 0) {
+    if (rrep.destAddress!.compareTo(_ownLabel!) == 0) {
       if (_verbose) log(TAG, '$_ownLabel is the destination (stop RREP)');
         _saveDestSequenceNumber(rrep.originAddress, rrep.sequenceNum);
 
         _aodvHelper.addEntryRoutingTable(rrep.originAddress, nextHop, hopRcv, rrep.sequenceNum, rrep.lifetime, null);
 
-        Data data = _dataMessage.pdu as Data;
+        Data data = _dataMessage!.pdu as Data;
         _send(_dataMessage, data.destAddress);
 
-        _timerFlushForwardRoute(rrep.originAddress, rrep.sequenceNum, rrep.lifetime);
+        _timerFlushForwardRoute(rrep.originAddress, rrep.sequenceNum, rrep.lifetime!);
     } else {
-      EntryRoutingTable destNext = _aodvHelper.getNextfromDest(rrep.destAddress);
+      EntryRoutingTable? destNext = _aodvHelper.getNextfromDest(rrep.destAddress);
       if (destNext == null) {
         throw AodvUnknownDestException('No destNext found in the routing Table for ${rrep.destAddress}');
       } else {
@@ -361,29 +361,29 @@ class AodvManager {
 
         _aodvHelper.addEntryRoutingTable(rrep.originAddress, nextHop, hopRcv, rrep.sequenceNum, rrep.lifetime, _addPrecursors(destNext.next));
 
-        _timerFlushForwardRoute(rrep.originAddress, rrep.sequenceNum, rrep.lifetime);
+        _timerFlushForwardRoute(rrep.originAddress, rrep.sequenceNum, rrep.lifetime!);
       }
     }
   }
 
   void _processRREP_GRATUITOUS(MessageAdHoc message) {
-    RREP rrep = RREP.fromJson(message.pdu as Map);
+    RREP rrep = RREP.fromJson((message.pdu as Map) as Map<String, dynamic>);
     int hopCount = rrep.incrementHopCount();
 
-    if (rrep.destAddress.compareTo(_ownLabel) == 0) {
+    if (rrep.destAddress!.compareTo(_ownLabel!) == 0) {
       if (_verbose) log(TAG, '$_ownLabel is the destination (stop RREP)');
 
-      _aodvHelper.addEntryRoutingTable(rrep.originAddress, message.header.label, hopCount, rrep.sequenceNum, rrep.lifetime, null);
+      _aodvHelper.addEntryRoutingTable(rrep.originAddress, message.header!.label, hopCount, rrep.sequenceNum, rrep.lifetime, null);
 
       // _timerFlushReverseRoute(rrep.originAddress, rrep.sequenceNum);
     } else {
-      EntryRoutingTable destNext = _aodvHelper.getNextfromDest(rrep.destAddress);
+      EntryRoutingTable? destNext = _aodvHelper.getNextfromDest(rrep.destAddress);
       if (destNext == null) {
         throw AodvUnknownDestException('No destNext found in the routing Table for ${rrep.destAddress}');
       } else {
         if (_verbose) log(TAG, 'Destination reachable via ${destNext.next}');
         
-        _aodvHelper.addEntryRoutingTable(rrep.originAddress, message.header.label, hopCount, rrep.sequenceNum, rrep.lifetime, _addPrecursors(destNext.next));
+        _aodvHelper.addEntryRoutingTable(rrep.originAddress, message.header!.label, hopCount, rrep.sequenceNum, rrep.lifetime, _addPrecursors(destNext.next));
 
         _timerFlushReverseRoute(rrep.originAddress, rrep.sequenceNum);
 
@@ -400,12 +400,12 @@ class AodvManager {
   }
 
   void _processRERR(MessageAdHoc message) {
-    RERR rerr = RERR.fromJson(message.pdu as Map);
-    String originateAddr = message.header.label;
+    RERR rerr = RERR.fromJson((message.pdu as Map) as Map<String, dynamic>);
+    String? originateAddr = message.header!.label;
 
     if (_verbose) log(TAG, 'Received RERR from $originateAddr -> Node ${rerr.unreachableDestAddress} is unreachable');
 
-    if (rerr.unreachableDestAddress.compareTo(_ownLabel) == 0) {
+    if (rerr.unreachableDestAddress!.compareTo(_ownLabel!) == 0) {
       if (_verbose) log(TAG, 'RERR received on the destination (stop forward)');
     } else if (_aodvHelper.containsDest(rerr.unreachableDestAddress)) {
       message.header = Header(
@@ -415,9 +415,9 @@ class AodvManager {
         mac: _ownMac
       );
         
-      List<String> precursors = _aodvHelper.getPrecursorsFromDest(rerr.unreachableDestAddress);
+      List<String?>? precursors = _aodvHelper.getPrecursorsFromDest(rerr.unreachableDestAddress);
       if (precursors != null) {
-        for (String precursor in precursors) {
+        for (String? precursor in precursors) {
           if (_verbose) log(TAG, ' Precursor: $precursor');
             _send(message, precursor);
         }
@@ -432,14 +432,14 @@ class AodvManager {
   }
 
   void _processData(MessageAdHoc message) {
-    Data data = Data.fromJson(message.pdu as Map);
+    Data data = Data.fromJson((message.pdu as Map) as Map<String, dynamic>);
 
-    if (_verbose) log(TAG, 'Data message received from: ${message.header.label}');
+    if (_verbose) log(TAG, 'Data message received from: ${message.header!.label}');
 
-    if (data.destAddress.compareTo(_ownLabel) == 0) {
-      if (_verbose) log(TAG, _ownLabel + ' is the destination (stop DATA message)');
+    if (data.destAddress!.compareTo(_ownLabel!) == 0) {
+      if (_verbose) log(TAG, _ownLabel! + ' is the destination (stop DATA message)');
 
-      Header header = message.header;
+      Header header = message.header!;
       AdHocDevice device = AdHocDevice(
         label: header.label,
         name: header.name,
@@ -449,13 +449,13 @@ class AodvManager {
 
       _eventCtrl.add(AdHocEvent(DatalinkConstants.DATA_RECEIVED, [device, data.payload]));
     } else {
-      EntryRoutingTable destNext = _aodvHelper.getNextfromDest(data.destAddress);
+      EntryRoutingTable? destNext = _aodvHelper.getNextfromDest(data.destAddress);
       if (destNext == null) {
         throw AodvUnknownDestException('No destNext found in the routing Table for ${data.destAddress}');
       } else {
         if (_verbose) log(TAG, 'Destination reachable via ${destNext.next}');
 
-        Header header = message.header;
+        Header header = message.header!;
         AdHocDevice device = AdHocDevice(
           label: header.label,
           name: header.name,
@@ -472,16 +472,16 @@ class AodvManager {
     }
   }
 
-  void _sendRRER(String brokenNodeAddress) {
+  void _sendRRER(String? brokenNodeAddress) {
     if (_aodvHelper.containsNext(brokenNodeAddress)) {
-      String dest = _aodvHelper.getDestFromNext(brokenNodeAddress);
-      if (dest.compareTo(_ownLabel) == 0) {
+      String dest = _aodvHelper.getDestFromNext(brokenNodeAddress)!;
+      if (dest.compareTo(_ownLabel!) == 0) {
         if (_verbose) log(TAG, 'RERR received on the destination (stop forward)');
       } else {  
         RERR rrer = RERR(type: AodvConstants.RERR, unreachableDestAddress: dest, unreachableDestSeqNum: _ownSequenceNum);
-        List<String> precursors = _aodvHelper.getPrecursorsFromDest(dest);
+        List<String?>? precursors = _aodvHelper.getPrecursorsFromDest(dest);
         if (precursors != null) {
-          for (String precursor in precursors) {
+          for (String? precursor in precursors) {
             if (_verbose) log(TAG, 'send RERR to $precursor');
             _send(
               MessageAdHoc(
@@ -503,8 +503,8 @@ class AodvManager {
     }
   }
 
-  void _sendRREP_GRATUITOUS(String senderAddr, RREQ rreq) {
-    EntryRoutingTable entry = _aodvHelper.getDestination(rreq.destAddress);
+  void _sendRREP_GRATUITOUS(String? senderAddr, RREQ rreq) {
+    EntryRoutingTable entry = _aodvHelper.getDestination(rreq.destAddress)!;
 
     entry.updatePrecursors(senderAddr);
 
@@ -541,7 +541,7 @@ class AodvManager {
 
     rrep = RREP(
       type: AodvConstants.RREP,
-      hopCount: entry.hop + 1, 
+      hopCount: entry.hop! + 1, 
       destAddress: rreq.originAddress, 
       sequenceNum: entry.destSeqNum, 
       originAddress:
@@ -564,12 +564,12 @@ class AodvManager {
     if (_verbose) log(TAG, 'Send RREP to ${rreq.originAddress}');
   }
 
-  void _timerFlushForwardRoute(String destAddress, int sequenceNum, int lifeTime) {
+  void _timerFlushForwardRoute(String? destAddress, int? sequenceNum, int lifeTime) {
     Timer(Duration(milliseconds: lifeTime),
       () {
         if (_verbose) log(TAG, 'Add timer for $destAddress - seq: $sequenceNum - lifeTime: $lifeTime');
 
-        int lastChanged = _aodvHelper.getDataPathFromAddress(destAddress);
+        int lastChanged = _aodvHelper.getDataPathFromAddress(destAddress)!;
         int difference = (DateTime.now().millisecond - lastChanged);
 
         if (lastChanged == 0) {
@@ -585,11 +585,11 @@ class AodvManager {
     );
   }
 
-  void _timerFlushReverseRoute(String originAddress, int sequenceNum) {
+  void _timerFlushReverseRoute(String? originAddress, int? sequenceNum) {
     Timer(Duration(milliseconds: AodvConstants.EXPIRED_TABLE),
       () {
         if (_verbose) log(TAG, 'Add timer for $originAddress - seq: $sequenceNum');
-        int lastChanged = _aodvHelper.getDataPathFromAddress(originAddress);
+        int lastChanged = _aodvHelper.getDataPathFromAddress(originAddress)!;
         int difference = (DateTime.now().millisecond - lastChanged);
         if (lastChanged == 0) {
           _aodvHelper.removeEntry(originAddress);
@@ -605,7 +605,7 @@ class AodvManager {
   }
 
   void _processAodvMsgReceived(MessageAdHoc message) {
-    switch (message.header.messageType) {
+    switch (message.header!.messageType) {
       case AodvConstants.RREQ:
         _processRREQ(message);
         _getNextSequenceNumber();
