@@ -16,12 +16,12 @@ import 'package:adhoc_plugin/src/network/datalinkmanager/wrapper_wifi.dart';
 
 
 class DataLinkManager {
-  List<WrapperNetwork> _wrappers;
-  HashMap<String, AdHocDevice> _mapAddressDevice;
-  StreamController<DiscoveryEvent> _discoveryCtrl;
-  StreamController<AdHocEvent> _eventCtrl;
+  late List<WrapperNetwork?> _wrappers;
+  HashMap<String, AdHocDevice>? _mapAddressDevice;
+  late StreamController<DiscoveryEvent> _discoveryCtrl;
+  late StreamController<AdHocEvent> _eventCtrl;
 
-  String label;
+  String? label;
 
   DataLinkManager(bool verbose, Config config) {
     this._mapAddressDevice = HashMap();
@@ -37,11 +37,11 @@ class DataLinkManager {
 
 /*------------------------------Getters & Setters-----------------------------*/
 
-  List<AdHocDevice> get directNeighbors {
-    List<AdHocDevice> neighbors = List.empty(growable: true);
+  List<AdHocDevice?> get directNeighbors {
+    List<AdHocDevice?> neighbors = List.empty(growable: true);
 
     for (int i = 0; i < NB_WRAPPERS; i++)
-      neighbors.addAll(_wrappers[i].directNeighbors);
+      neighbors.addAll(_wrappers[i]!.directNeighbors);
 
     return neighbors;
   }
@@ -54,33 +54,33 @@ class DataLinkManager {
 
   int checkState() {
     int enabled = 0;
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
         enabled++;
 
     return enabled;
   }
 
   void enable(int duration, int type) {
-    _wrappers[type].enable(duration);
+    _wrappers[type]!.enable(duration);
   }
 
   void enableAll() {
-    for (WrapperNetwork wrapper in _wrappers)
-      enable(3600, wrapper.type);
+    for (WrapperNetwork? wrapper in _wrappers)
+      enable(3600, wrapper!.type!);
   }
 
   void disable(int type) {
-    if (_wrappers[type].enabled) {
-      _wrappers[type].stopListening();
-      _wrappers[type].disable();
+    if (_wrappers[type]!.enabled!) {
+      _wrappers[type]!.stopListening();
+      _wrappers[type]!.disable();
     }
   }
 
   void disableAll() {
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
-        disable(wrapper.type);
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
+        disable(wrapper.type!);
   }
 
   void discovery() {
@@ -91,8 +91,8 @@ class DataLinkManager {
     if (enabled == _wrappers.length) {
       _discovery();
     } else {
-      for (WrapperNetwork wrapper in _wrappers) {
-        if (wrapper.enabled) {
+      for (WrapperNetwork? wrapper in _wrappers) {
+        if (wrapper!.enabled!) {
           wrapper.discovery();
         }
       }
@@ -102,55 +102,55 @@ class DataLinkManager {
   Future<void> connect(int attempts, AdHocDevice device) async {
     switch (device.type) {
       case WIFI:
-        await _wrappers[WIFI].connect(attempts, device);
+        await _wrappers[WIFI]!.connect(attempts, device);
         break;
 
       case BLE:
-        await _wrappers[BLE].connect(attempts, device);
+        await _wrappers[BLE]!.connect(attempts, device);
         break;
     }
   }
 
   void stopListening() {
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
         wrapper.stopListening();
   }
 
   void removeGroup() {
-    WrapperWifi wrapperWifi = _wrappers[WIFI];
-    if (wrapperWifi.enabled) {
+    WrapperWifi wrapperWifi = _wrappers[WIFI] as WrapperWifi;
+    if (wrapperWifi.enabled!) {
       wrapperWifi.removeGroup();
     } else {
       throw DeviceFailureException("Wifi is not enabled");
     }
   }
 
-  bool isWifiGroupOwner() {
-    WrapperWifi wrapperWifi = _wrappers[WIFI];
-    if (wrapperWifi.enabled) {
+  bool? isWifiGroupOwner() {
+    WrapperWifi wrapperWifi = _wrappers[WIFI] as WrapperWifi;
+    if (wrapperWifi.enabled!) {
       return wrapperWifi.isWifiGroupOwner();
     } else {
       throw DeviceFailureException("Wifi is not enabled");
     }
   }
 
-  void sendMessage(MessageAdHoc message, String address) {
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
+  void sendMessage(MessageAdHoc? message, String? address) {
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
         wrapper.sendMessage(message, address);
   }
 
   void broadcast(MessageAdHoc message) {
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
         wrapper.broadcast(message);
   }
 
   Future<bool> broadcastObject(Object object) async {
     bool sent = false;
-    for (WrapperNetwork wrapper in _wrappers) {
-      if (wrapper.enabled) {
+    for (WrapperNetwork? wrapper in _wrappers) {
+      if (wrapper!.enabled!) {
         Header header = Header(
           messageType: BROADCAST,
           label: label,
@@ -166,16 +166,16 @@ class DataLinkManager {
     return sent;
   }
 
-  void broadcastExcept(MessageAdHoc message, String excludedAddress) {
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
+  void broadcastExcept(MessageAdHoc message, String? excludedAddress) {
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
         wrapper.broadcastExcept(message, excludedAddress);
   }
 
-  Future<bool> broadcastObjectExcept(Object object, String excludedAddress) async {
+  Future<bool> broadcastObjectExcept(Object object, String? excludedAddress) async {
     bool sent = false;
-    for (WrapperNetwork wrapper in _wrappers) {
-      if (wrapper.enabled) {
+    for (WrapperNetwork? wrapper in _wrappers) {
+      if (wrapper!.enabled!) {
         Header header = Header(
           messageType: BROADCAST,
           label: label,
@@ -191,41 +191,41 @@ class DataLinkManager {
     return sent;
   }
 
-  Future<HashMap<String, AdHocDevice>> getPaired() async {
-    if (_wrappers[BLE].enabled)
-      return await _wrappers[BLE].getPaired();
+  Future<HashMap<String?, AdHocDevice>?> getPaired() async {
+    if (_wrappers[BLE]!.enabled!)
+      return await _wrappers[BLE]!.getPaired();
     return null;
   }
 
-  bool isDirectNeighbors(String address) {
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled && wrapper.isDirectNeighbors(address))
+  bool isDirectNeighbors(String? address) {
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled! && wrapper.isDirectNeighbors(address))
         return true;
     return false;
   }
 
-  List<AdHocDevice> getDirectNeighbors() {
-    List<AdHocDevice> devices = List.empty(growable: true);
+  List<AdHocDevice?> getDirectNeighbors() {
+    List<AdHocDevice?> devices = List.empty(growable: true);
 
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
         devices.addAll(wrapper.directNeighbors);
 
     return devices;
   }
 
-  bool isEnabled(int type) => _wrappers[type].enabled;
+  bool? isEnabled(int type) => _wrappers[type]!.enabled;
 
-  Future<String> getAdapterName(int type) async {
-    if (_wrappers[type].enabled)
-      return await _wrappers[type].getAdapterName();
+  Future<String?> getAdapterName(int type) async {
+    if (_wrappers[type]!.enabled!)
+      return await _wrappers[type]!.getAdapterName();
     return null;
   }
 
-  Future<HashMap<int, String>> getActifAdapterNames() async {
-    HashMap<int, String> adapterNames = HashMap();
-    for (WrapperNetwork wrapper in _wrappers) {
-      String name = await getAdapterName(wrapper.type);
+  Future<HashMap<int?, String>> getActifAdapterNames() async {
+    HashMap<int?, String> adapterNames = HashMap();
+    for (WrapperNetwork? wrapper in _wrappers) {
+      String? name = await getAdapterName(wrapper!.type!);
       if (name != null)
         adapterNames.putIfAbsent(wrapper.type, () => name);
     }
@@ -233,9 +233,9 @@ class DataLinkManager {
     return adapterNames;
   }
 
-  Future<bool> updateAdapterName(int type, String newName) async {
-    if (_wrappers[type].enabled) {
-      return await _wrappers[type].updateDeviceName(newName);
+  Future<bool?> updateAdapterName(int type, String newName) async {
+    if (_wrappers[type]!.enabled!) {
+      return await _wrappers[type]!.updateDeviceName(newName);
     } else {
       throw DeviceFailureException(
         _typeString(type) + ' adapter is not enabled'
@@ -244,42 +244,42 @@ class DataLinkManager {
   }
 
   void resetAdapterName(int type) {
-    if (_wrappers[type].enabled) {
-      _wrappers[type].resetDeviceName();
+    if (_wrappers[type]!.enabled!) {
+      _wrappers[type]!.resetDeviceName();
     } else {
       throw DeviceFailureException(_typeString(type) + ' adapter is not enabled');
     }
   }
 
   void disconnectAll() {
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
         wrapper.disconnectAll();
   }
 
-  void disconnect(String remoteDest) {
-    for (WrapperNetwork wrapper in _wrappers)
-      if (wrapper.enabled)
+  void disconnect(String? remoteDest) {
+    for (WrapperNetwork? wrapper in _wrappers)
+      if (wrapper!.enabled!)
         wrapper.disconnect(remoteDest);
   }
 
 /*------------------------------Private methods-------------------------------*/
 
   void _initialize() {
-    _wrappers[BLE].eventStream.listen((event) => _eventCtrl.add(event));
-    _wrappers[BLE].discoveryStream.listen((event) => _discoveryCtrl.add(event));
-    _wrappers[WIFI].eventStream.listen((event) => _eventCtrl.add(event));
-    _wrappers[WIFI].discoveryStream.listen((event) => _discoveryCtrl.add(event));
+    _wrappers[BLE]!.eventStream.listen((event) => _eventCtrl.add(event));
+    _wrappers[BLE]!.discoveryStream.listen((event) => _discoveryCtrl.add(event));
+    _wrappers[WIFI]!.eventStream.listen((event) => _eventCtrl.add(event));
+    _wrappers[WIFI]!.discoveryStream.listen((event) => _discoveryCtrl.add(event));
   }
 
   void _discovery() {
-    for (WrapperNetwork wrapper in _wrappers)
-      wrapper.discovery();
+    for (WrapperNetwork? wrapper in _wrappers)
+      wrapper!.discovery();
 
     Timer.periodic(Duration(milliseconds: POOLING_DISCOVERY), (Timer timer) {
       bool finished = true;
-      for (WrapperNetwork wrapper in _wrappers) {
-        if (!wrapper.discoveryCompleted) {
+      for (WrapperNetwork? wrapper in _wrappers) {
+        if (!wrapper!.discoveryCompleted) {
           finished = false;
           break;
         }
@@ -289,8 +289,8 @@ class DataLinkManager {
         timer.cancel();
     });
 
-    for (WrapperNetwork wrapper in _wrappers)
-      wrapper.discoveryCompleted = false;
+    for (WrapperNetwork? wrapper in _wrappers)
+      wrapper!.discoveryCompleted = false;
   }
 
   String _typeString(int type) {
