@@ -4,12 +4,12 @@ import 'package:pointycastle/pointycastle.dart';
 
 
 class Certificate {
-  String _owner;
-  String _issuer;
-  DateTime _validity;
+  String? _owner;
+  String? _issuer;
+  DateTime? _validity;
   RSAPublicKey _key;
 
-  Uint8List signature;
+  Uint8List? signature;
 
   Certificate(this._owner, this._issuer, this._key) {
     this.signature = Uint8List(1);
@@ -17,11 +17,11 @@ class Certificate {
 
 /*------------------------------Getters & Setters-----------------------------*/
 
-  String get owner => _owner;
+  String? get owner => _owner;
 
-  String get issuer => _issuer;
+  String? get issuer => _issuer;
 
-  DateTime get validity => _validity;
+  DateTime? get validity => _validity;
 
   RSAPublicKey get key => _key;
 }
