@@ -1,7 +1,7 @@
 class GroupData {
-  String leader;
-  int groupId;
-  Object data;
+  String? leader;
+  int? groupId;
+  Object? data;
 
   GroupData(this.leader, this.groupId, this.data);
 }
