@@ -3,7 +3,7 @@ import 'dart:typed_data';
 
 import 'package:adhoc_plugin/src/datalink/utils/msg_adhoc.dart';
 
-String checkString(String string) => string == null ? '' : string;
+String checkString(String? string) => string == null ? '' : string;
 
 void log(final String tag, final String message) => print(tag + ': ' + message);
 
